+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Version 0.0.7
+
+- Rust client examples in programs/multisig-lite/examples.
+- Run Rust client examples in CI.
+- Run anchor test in CI.
+- Add preliminary solana-program-test based functional tests.
+
 # Version 0.0.6
 
 - Deployed to the [devnet].